added readme and moved things around in file stutcure
